+++ integration-test/src/test/resources/trino-ci-testset/testsets/jdbc-mysql/bugs/00004_query_pushdown.sql
