+++ integration-test/src/test/resources/trino-ci-testset/testsets/jdbc-postgresql/bugs/00003_join_pushdown.txt
