+++ integration-test/src/test/resources/trino-ci-testset/testsets/jdbc-postgresql/bugs/00004_query_pushdown.txt
