+++ trino-connector/src/main/java/com/datastrato/gravitino/trino/connector/GravitinoDataSourceProvider.java
@@ -12,10 +12,7 @@ import io.trino.spi.connector.ConnectorSplit;
 import io.trino.spi.connector.ConnectorTableHandle;
 import io.trino.spi.connector.ConnectorTransactionHandle;
 import io.trino.spi.connector.DynamicFilter;
-import io.trino.spi.predicate.TupleDomain;
 import java.util.List;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
 
 /** This class provides a ConnectorPageSource for trino read data from internal connector. */
 public class GravitinoDataSourceProvider implements ConnectorPageSourceProvider {
@@ -41,38 +38,5 @@ public class GravitinoDataSourceProvider implements ConnectorPageSourceProvider 
         GravitinoHandle.unWrap(table),
         GravitinoHandle.unWrap(columns),
         new GravitinoDynamicFilter(dynamicFilter));
-  }
-
-  static class GravitinoDynamicFilter implements DynamicFilter {
-    private final DynamicFilter delegate;
-
-    GravitinoDynamicFilter(DynamicFilter dynamicFilter) {
-      delegate = dynamicFilter;
-    }
-
-    @Override
-    public Set<ColumnHandle> getColumnsCovered() {
-      return delegate.getColumnsCovered();
-    }
-
-    @Override
-    public CompletableFuture<?> isBlocked() {
-      return delegate.isBlocked();
-    }
-
-    @Override
-    public boolean isComplete() {
-      return delegate.isComplete();
-    }
-
-    @Override
-    public boolean isAwaitable() {
-      return delegate.isAwaitable();
-    }
-
-    @Override
-    public TupleDomain<ColumnHandle> getCurrentPredicate() {
-      return delegate.getCurrentPredicate().transformKeys(GravitinoHandle::unWrap);
-    }
   }
 }
